requirements.txt
torch
transformers
numpy
pytest
nltk
pyyaml